personal_library_manager_with_mongodb/pyproject.toml
[project]
name = "personal-library-manager-with-mongodb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pymongo>=4.11.3",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
]
